requirements.txt
flask==3.0.2
flask-cors==4.0.0
pymongo==4.6.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3 